Add test for resetObservation
